+++ database/migrations/2019_07_27_104941_create_table_transportation.php
@@ -13,24 +13,26 @@ class CreateTableTransportation extends Migration
      */
     public function up()
     {
-      Schema::create('transportation', function (Blueprint $table) {
-        $table->bigIncrements('id');
-        $table->unsignedInteger('vehicleTypeId');
-        $table->foreign('vehicleTypeId')
-              ->references('id')
-              ->on('vehicle_type')
-              ->onDelete('cascade')
-              ->onUpdate('cascade');
-        $table->string('origin', 150);
-        $table->string('destination', 150);
-        $table->string('code', 100)->nullable();
-        $table->string('fleet', 180);
-        $table->string('receipt', 255)->nullable();
-        $table->text('remark')->nullable();
-        $table->timestamps();
-        $table->text('created_by')->nullable();
-        $table->text('updated_by')->nullable();
-      });
+        if(!Schema::hasTable('transportation')) {
+          Schema::create('transportation', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->unsignedInteger('vehicleTypeId');
+            $table->foreign('vehicleTypeId')
+                  ->references('id')
+                  ->on('vehicle_type')
+                  ->onDelete('cascade')
+                  ->onUpdate('cascade');
+            $table->string('origin', 150);
+            $table->string('destination', 150);
+            $table->string('code', 100)->nullable();
+            $table->string('fleet', 180);
+            $table->string('receipt', 255)->nullable();
+            $table->text('remark')->nullable();
+            $table->timestamps();
+            $table->text('created_by')->nullable();
+            $table->text('updated_by')->nullable();
+          });
+        }
     }
 
     /**
@@ -40,6 +42,8 @@ class CreateTableTransportation extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('transportation');
+        if(Schema::hasTable('transportation')) {
+          Schema::dropIfExists('transportation');
+        }
     }
 }
